:recycle: Refactor: remove unused code
