+++ src/components/QueryEditor/DatabricksQueryEditor.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { QueryEditorProps } from '@grafana/data';
 import {SqlQueryEditor, SQLOptions, QueryHeaderProps, SQLQuery} from 'components/grafana-sql/src';
 
-import { DatabricksDatasource } from './datasource';
+import { DatabricksDatasource } from '../../datasource';
 
 const queryHeaderProps: Pick<QueryHeaderProps, 'dialect'> = { dialect: 'other' };
