+++ src/components/Suggestions/sqlUtil.ts
