+++ src/types.ts
@@ -1,15 +1,4 @@
 import { SQLOptions } from 'components/grafana-sql/src';
-import {DataQuery} from '@grafana/data';
-
-interface QuerySettings {
-  convertLongToWide: boolean
-  fillMode?: number
-  fillValue?: number
-}
-export interface MyQuery extends DataQuery {
-  rawSqlQuery?: string;
-  querySettings: QuerySettings;
-}
 
 /**
  * These are options configured for each DataSource instance.
@@ -30,18 +19,6 @@ export interface DatabricksDataSourceOptions extends SQLOptions {
 export interface DatabricksSecureJsonData {
   token?: string;
   clientSecret?: string;
-}
-
-export interface Column {
-  name: string
-  type: string
-}
-
-export interface Suggestions {
-  catalogs: string[]
-  schemas: string[]
-  tables: string[]
-  columns: Column[]
 }
 
 export type ColumnResponse = {
